requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For reading Excel files
sqlalchemy>=2.0.0
alembic>=1.13.0  # Database migrations
psycopg2-binary>=2.9.0  # PostgreSQL adapter
pydantic>=2.0.0
python-dotenv>=1.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data processing
python-dateutil>=2.8.0

# Redis & Job Queue
redis>=5.0.0
arq>=0.25.0  # Async job queue

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-exporter-prometheus==0.60b1
structlog>=24.1.0  # Structured JSON logging

# ML
scikit-learn>=1.3.0
torch>=2.1.0
shap>=0.44.0  # For explainability

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
locust>=2.17.0  # Load testing

# Development
black>=23.0.0
mypy>=1.5.0


